# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.1",
    "title": "Your pet API docs",
    "description": "Endpoints documentation for your pet API"
  },
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json", "multipart/form-data"],
...
  "paths": {
...
    "/api/pets": {
      "get": {
        "summary": "Get user's pets",
        "description": "Get user's pets. Requires authentication token.",
        "tags": ["Pets"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIi..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "401": {
            "description": "Not authorized",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      },
      "post": {
        "summary": "Add user's pet",
        "description": "Add user's pet. Requires authentication token.",
        "tags": ["Pets"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIi..."
            }
          }
        ],
        "requestBody": {
          "description": "New pet's data object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetRequest"
              }
            }
          }
        },

        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {}
          },
          "400": {
            "description": "Bad request (invalid request body)",
            "content": {}
          },
          "401": {
            "description": "Not authorized",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      }
    },
    "/api/pets/{petId}": {
      "delete": {
        "summary": "Delete user's pet",
        "description": "Delete user's pets. Requires authentication token.",
        "tags": ["Pets"],
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer token",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIi..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {}
          },
          "401": {
            "description": "No authorization",
            "content": {}
          },
          "404": {
            "description": "Not found",
            "content": {}
          },
          "500": {
            "description": "Server error",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePetRequest": {
        "type": "object",
        "required": ["name", "birthday", "breed"],
        "properties": {
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "User's avatar"
          },
          "name": {
            "type": "string",
            "description": "Pet's name",
            "minLength": 2,
            "maxLength": 16,
            "pattern": "/^[a-zA-Z ]+$/",
            "example": "Napoleon"
          },
          "birthday": {
            "type": "string",
            "description": "Pet's birthday",
            "format": "date",
            "example": "2001-01-01"
          },
          "breed": {
            "type": "string",
            "description": "Pet's breed",
            "minLength": 2,
            "maxLength": 16,
            "pattern": "/^[a-zA-Z ]+$/",
            "example": "buldog"
          },
          "comments": {
            "type": "string",
            "description": "Comment about pet",
            "minLength": 8,
            "maxLength": 120,
            "example": "This is a good dog"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}